main.go: adapt script to use stdin and out for reading and deploying
the original script reads a file from disk and then writes this to disk.
this commit adapts the code in the way that it expects the template to
be piped into stdin and uses the "echo $template | kubectl apply -f -'
command to apply the template. By doing so, we omit having decrypted
secrets flying around somewhere on disk at any given point in time.