ops: always use GOOGLE_APPLICATION_CREDENTIALS from env
this is the de-facto standard and I don't see why this should
change here or be passed into the program any differently.
docker-entrypoint.sh now writes the JSON string into a file
with the right permissions and sets the environment variable
GOOGLE_APPLICATION_CREDENTIALS accordingly. This allows both
drone-gke as well as sops to use the same credentials without
any additional configuration.